backend/requirements.txt
# Web framework
fastapi>=0.115
uvicorn[standard]>=0.30

# Configuration & environment
python-dotenv>=1.0
pydantic>=2.0

# Form / file upload support
python-multipart>=0.0.9

# LLM providers
openai>=1.0
anthropic>=0.25

# Frame generation
manim
cairosvg
lxml
requests
playwright

# Video pipeline
moviepy
imageio-ffmpeg
Pillow>=10.0

# Text-to-speech
gtts

# Auth
google-auth>=2.0
PyJWT>=2.8
bcrypt>=4.0
email-validator>=2.0

# Rate limiting
slowapi>=0.1.9

# Deep research
tavily-python>=0.3.0
pdfplumber>=0.10.0
python-pptx>=0.6.21
chromadb>=0.5.0

# Observability
structlog>=24.0
prometheus-client>=0.20.0
